ml: more examples added
git-svn-id: http://k-framework.googlecode.com/svn/trunk@1669 82a86efa-2583-11df-8ea0-49ba94de29ab

Former-commit-id: b2cc456886bd121c31f58395af2a2d5913716d14

+++ examples/matchC/bin/matchC.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+#set -x
 
 K_ROOT_DIR=~/k-framework
 ML_ROOT_DIR=${K_ROOT_DIR}/examples/matchC
@@ -28,8 +29,21 @@ MFLAGS=-no-banner
 TMP_OUT=.tmp_out_file
 TMP_ERR=.tmp_err_file
 
-if [ ! "$1" ]; then
-  echo "mlc: no input file"
+while getopts 'eco:' OPTION; do
+  case $OPTION in
+    c) COMPILE_FLAG="-c"
+       ;;
+    e) EXEC_FLAG="-e"
+       ;;
+    o) OUT_FLAG='-o'
+       ML_PROG="$OPTARG"
+       ;;
+    ?) ;; 
+  esac
+done
+shift $(($OPTIND - 1))
+if [ -z "$1" ]; then
+  echo "matchC: no input file"
   exit 1
 fi
 if [ ! -e $1 ]; then
@@ -41,11 +55,18 @@ PROG=`basename $1 ${LANG_SUFFIX}`
 PROG_DIR=`dirname $1`
 PROG_MODULE=PROG
 PROG_MACRO=prog
-MAUDE_PROG=${PROG}.maude
-COMPILED_PROG=prog-compiled.maude
-ML_PROG=${PROG}-ml.maude
+MAUDE_PROG=${PROG_MACRO}.maude
+COMPILED_PROG=${PROG_MACRO}-compiled.maude
+if [ -z "${OUT_FLAG}" ]; then
+  ML_PROG=${PROG}.maude
+fi
+if [ -z "${EXEC_FLAG}" ]; then
+  ML_OP=check
+else
+  ML_OP=run
+fi
 
-PROG_BODY=`grep -v '^#include' $1 | grep -v 'printf(' | ${JVM} ${JFLAGS} ${PARSER_MAIN}`
+PROG_BODY=`grep -v '^#include' $1 | ${JVM} ${JFLAGS} ${PARSER_MAIN}`
 if [ "$?" -ne 0 ]; then exit $?; fi
 
 echo -e "
@@ -53,31 +74,41 @@ load ${ML_LIB_DIR}/ml-prelude.maude\n\
 load ${LANG_SEMANTICS_DIR}/${LANG_NAME}-syntax.maude\n\
 load ${LANG_SEMANTICS_DIR}/${LANG_NAME}-flat-compiled.maude\n\
 mod ${PROG_MODULE} is\n\
-including ${LANG_MODULE}-SYNTAX + MATHEMATICAL-DOMAIN-BUILTIN-MODULE + LIST-HP + BINARY-TREE-HP .\n\
-${PROG_BODY}\n\
-endm\n\
+including ${LANG_MODULE}-SYNTAX + MATHEMATICAL-DOMAIN-BUILTIN-MODULE + LIST-HP + BINARY-TREE-HP .
 " >${MAUDE_PROG}
+echo "${PROG_BODY}" >>${MAUDE_PROG}
+echo -e "
+endm\n\
+" >>${MAUDE_PROG}
 
 ${KC} ${KFLAGS} ${MAUDE_PROG} ${LANG_MODULE} ${PROG_MODULE} ${PROG_MACRO} >${TMP_ERR}
 if [ "$?" -ne 0 ]; then ERR=$?; cat ${TMP_ERR}; rm ${TMP_ERR}; exit ${ERR}; fi
 
-UNWRAPED_BODY=`${JVM} ${JFLAGS} ${UNWRAP_MAIN} <${COMPILED_PROG} | grep -v '^load "list"'`
+UNWRAPED_BODY=`${JVM} ${JFLAGS} ${UNWRAP_MAIN} <${COMPILED_PROG} | grep -v '^load'`
 if [ "$?" -ne 0 ]; then exit $?; fi
 
 echo -e "
 load ${ML_LIB_DIR}/ml-prelude.maude\n\
 load ${LANG_SEMANTICS_DIR}/${LANG_NAME}-uw.maude\n\
-load ${ML_LIB_DIR}/fol.maude\n\
-${UNWRAPED_BODY}\n\
-mod TEST is inc ${LANG_MODULE}-${PROG_MACRO} + FOL= . endm\n\
-set print attribute on .\n\
-rew check('prog) .\n\
-q\n\
+load ${ML_LIB_DIR}/fol.maude
 " >${ML_PROG}
+echo "${UNWRAPED_BODY}" >>${ML_PROG}
+echo -e "
+mod TEST is inc ${LANG_MODULE}-${PROG_MACRO} + FOL= . endm\n\
+set print attribute on .\n
+rew ${ML_OP}('prog) .\n
+q\n\
+" >>${ML_PROG}
 
-${MAUDE} ${MFLAGS} ${ML_PROG} >${TMP_OUT} 2>${TMP_ERR}
-if [ "$?" -ne 0 ]; then ERR=$?; cat ${TMP_ERR}; rm ${TMP_ERR}; exit ${ERR}; fi
+if [ -z "${COMPILE_FLAG}" ]; then
+  ${MAUDE} ${MFLAGS} ${ML_PROG} >${TMP_OUT} 2>${TMP_ERR}
+  if [ "$?" -ne 0 ]; then ERR=$?; cat ${TMP_ERR}; rm ${TMP_ERR}; exit ${ERR}; fi
 
-sed -e '1,2d' -e '$d' ${TMP_OUT}
-#rm -f ${TMP_OUT} ${TMP_ERR} ${MAUDE_PROG} ${COMPILED_PROG} ${ML_PROG}
+  sed -e '1,2d' -e '$d' ${TMP_OUT}
+fi
+
+rm -f ${TMP_OUT} ${TMP_ERR} ${MAUDE_PROG} ${COMPILED_PROG}
+if [ -z "${COMPILE_FLAG}" -a -z "${OUT_FLAG}" ]; then
+  rm -f ${ML_PROG}
+fi
